Moved healthcheck to start.sh
Apparently the original health_check only sets the container to 'unhealthy' and exit 1 does not actually exit the Docker with exit code 1. To fix this, we just use a while true loop in the start.sh, which also runs when the Docker start. Eventhough this feels like kinda  a workaround, it seems to work fine.  Also 2 new environment variables have been added, HEALTH_CHECK_INTERVAL and HEALTH_CHECK_SILENT. Refer to the Readme to see what these do. The names speak for themselves also actually. Closes #25